move sparse package to top level
