+++ hashprobe/wyhash/wyhash_amd64.go
@@ -2,7 +2,7 @@
 
 package wyhash
 
-import "github.com/segmentio/parquet-go/hashprobe/sparse"
+import "github.com/segmentio/parquet-go/sparse"
 
 //go:noescape
 func MultiHashUint32Array(hashes []uintptr, values sparse.Uint32Array, seed uintptr)
