+++ hashprobe/wyhash/wyhash_purego.go
@@ -2,7 +2,7 @@
 
 package wyhash
 
-import "github.com/segmentio/parquet-go/hashprobe/sparse"
+import "github.com/segmentio/parquet-go/sparse"
 
 func MultiHashUint32Array(hashes []uintptr, values sparse.Uint32Array, seed uintptr) {
 	for i := range hashes {
